+++ CHANGELOG.md
@@ -10,6 +10,7 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 - Sync missing order data from gateway.
 ### Fixed
 - Fix: Duplicate ok callback triggers order status change.
+- Change order status to cancelled if callback fail is received for an existing order.
 
 ## [3.7.6]
 ### Added